requirements.txt
django
django-colorfield
django-ckeditor
